Removed inactive code. Refactored package structure.
