+++ SynonymCreator/src/main/scala/no/nrk/samnorsk/synonymmapper/SentenceSegmenter.scala
@@ -1,4 +1,4 @@
-package no.nrk.samnorsk.wikiextractor
+package no.nrk.samnorsk.synonymmapper
 
 import epic.preprocess.MLSentenceSegmenter
