+++ SynonymCreator/src/main/scala/no/nrk/samnorsk/synonymmapper/SynonymMapper.scala
@@ -5,15 +5,10 @@ import java.lang.Thread.UncaughtExceptionHandler
 import java.util.zip.GZIPInputStream
 
 import com.typesafe.scalalogging.slf4j.LazyLogging
-import no.nrk.samnorsk.util.IOUtils
-import no.nrk.samnorsk.wikiextractor._
 import scopt.RenderingMode.TwoColumns
 
-import scala.collection.mutable
 import scala.collection.mutable.ListBuffer
 import scala.io.{Codec, Source}
-
-// TODO expansion/reduction, filtering, output format
 
 case class DumpDescription(dump: File, fromLanguage: Language, toLanguage: Language)
 
@@ -30,19 +25,6 @@ object SynonymMapper extends LazyLogging {
   case class Mapping[A, B](source: A, target: B)
 
   type StringMapping = Mapping[String, String]
-
-  case class WordAndFrequency(word: String, mappingFrequency: Int)
-
-  case class SynonymLine(synonyms: Seq[String], canonicalForm: String) {
-
-    def getReductionSynonyms: String = {
-      synonyms.mkString("", ",", " => ") + canonicalForm
-    }
-
-    def getExpansionSynonyms: String = {
-      (synonyms :+ canonicalForm).mkString(",")
-    }
-  }
 
   def processDumps(dumps: Seq[DumpDescription], mapper: TranslationMapper[String, String], limit: Option[Int] = None,
                    filterParams: CounterFilterParams = CounterFilterParams()): Seq[TranslationCounter[String, String]] = {
@@ -64,89 +46,6 @@ object SynonymMapper extends LazyLogging {
     procs.foreach(_.iter.source.close())
 
     procs.map(_.counter)
-  }
-
-  def createSynonymsFromFrequencies(frequencyMap: Map[StringMapping, Int]): Seq[SynonymLine] = {
-
-    def aggregateFrequencies(frequencyMap: Map[StringMapping, Int], cutoff: Int = 5): Map[String, Seq[WordAndFrequency]] = {
-
-      val createInsert = (map: mutable.Map[String, Seq[WordAndFrequency]], mapping: StringMapping, frequency: Int) =>
-        map += (mapping.source -> (map.getOrElse(mapping.source, Seq()) :+ WordAndFrequency(mapping.target, frequency)))
-
-      frequencyMap
-        .filter(x => x._2 > cutoff)
-        .foldLeft(mutable.Map.empty[String, Seq[WordAndFrequency]]) { (agg, mappingAndFreq) => createInsert(agg, mappingAndFreq._1, mappingAndFreq._2) }
-        .toMap
-    }
-
-    val reverseFreqMap = frequencyMap.toSeq
-      .map { case (mapping, frequency) => (Mapping(mapping.target, mapping.source), frequency) }
-      .toMap
-
-    val reverseAggregatedMap = aggregateFrequencies(reverseFreqMap)
-
-    def isMostFrequentTranslation(candidate: String, source: String) = {
-      val wordsAndFreqForCandidate = reverseAggregatedMap.getOrElse(candidate, Seq()).sortBy(_.mappingFrequency).reverse
-      wordsAndFreqForCandidate.head.word == source
-    }
-
-    /**
-      * Only keep the synonyms which most frequently map to the source word.
-      */
-    val filteredAggregationMap: Map[String, Seq[WordAndFrequency]] = aggregateFrequencies(frequencyMap)
-      .map { case (sourceword, wordsAndFrequencies) =>
-        val filtered = wordsAndFrequencies.filter(x => isMostFrequentTranslation(x.word, sourceword))
-        (sourceword, filtered)
-      }
-
-    /**
-      * Keep the synonym candidates which are frequent enough, compared to the most frequent mapping. Don't allow candidate words to
-      * appear both on the lhs and rhs.
-      */
-    def getSynonymsAboveRelativeFreqThreshold(candidates: Seq[WordAndFrequency], highestTranslationFrequency: Int, threshold: Double = 0.3) = {
-      candidates
-        .filter(x => x.mappingFrequency > highestTranslationFrequency * threshold)
-        .filter(x => filteredAggregationMap.getOrElse(x.word, Seq()).isEmpty)
-    }
-
-    def getReverseMappedSynonyms(sourceWord: String) = {
-      val synonymCandidates = reverseAggregatedMap.getOrElse(sourceWord, Seq())
-      if (synonymCandidates.nonEmpty) {
-        val highestFrequency = synonymCandidates
-          .sortBy(_.mappingFrequency)
-          .last
-          .mappingFrequency
-
-        getSynonymsAboveRelativeFreqThreshold(synonymCandidates, highestFrequency)
-      } else {
-        Seq()
-      }
-    }
-
-    val synonyms: Map[String, Seq[WordAndFrequency]] = filteredAggregationMap
-      .filter(x => x._2.nonEmpty)
-      .map { case (sourceWord, wordsAndFreqs) =>
-        val sorted = wordsAndFreqs.sortBy(_.mappingFrequency).reverse
-        val highestFrequency = sorted
-          .head
-          .mappingFrequency
-
-        val mappingsWithHighFrequencies = getSynonymsAboveRelativeFreqThreshold(sorted, highestFrequency)
-
-        (sourceWord, (mappingsWithHighFrequencies ++ getReverseMappedSynonyms(sourceWord)).distinct)
-      }
-
-    synonyms.map(x => SynonymLine(x._2.map(_.word), x._1))
-      .map(synonymLine => SynonymLine(synonymLine.synonyms.filter(x => x != synonymLine.canonicalForm).distinct, synonymLine.canonicalForm))
-      .filter(synonymLine => synonymLine.synonyms.nonEmpty)
-      .toSeq
-  }
-
-  def writeSynonyms(synonymLines: Seq[SynonymLine], output: File, expansion: Boolean): Unit = {
-    IOUtils.wipeAndCreateNewFile(output)
-    synonymLines.map(x => if (expansion) x.getExpansionSynonyms else x.getReductionSynonyms)
-      .grouped(1000)
-      .foreach(lines => IOUtils.writeOutput(lines, output))
   }
 
   def main(args: Array[String]): Unit = {
