+++ SynonymCreator/src/main/scala/no/nrk/samnorsk/synonymmapper/TranslationCounter.scala
@@ -1,4 +1,4 @@
-package no.nrk.samnorsk.wikiextractor
+package no.nrk.samnorsk.synonymmapper
 
 import java.io.{File, FileWriter}
